minimal support for non-default LSPs
