+++ editors/code/src/settings.ts
@@ -1,0 +1,13 @@
+import * as vscode from "vscode";
+
+
+export interface ISettings {
+    binaryPath: string,
+}
+
+export function load(myPluginId: string): ISettings {
+    let configuration = vscode.workspace.getConfiguration(myPluginId);
+    return {
+        binaryPath: configuration.get<string>("binaryPath", "")
+    }
+}
